backend/requirements.txt
fastapi
uvicorn
supabase
pandas
scikit-learn
xgboost
python-dotenv
requests
yfinance
pandas_ta
